Updating the pulling script

+++ files/pull_config.sh
@@ -1,7 +1,13 @@
 #!/bin/bash
 
+[[ -z "$1" ]] && echo "usage: $0 GIT_URL" && exit 1
+
 cd $(dirname $0)
-git init
-git remote add origin $1
-git fetch --all
-git reset --hard origin/master
+
+[[ ! -d .git ]] && git init
+
+ORIGIN=$(git remote -v | grep 'origin' | grep '(fetch)' | awk '{print $2}')
+[[ "$ORIGIN" != $1 ]] && git remote add -f origin $1
+
+FETCH=$(git remote update origin)
+[[ "$FETCH" != "Fetching origin" ]] && git reset --hard origin/master
